[FIX] l10n_mx_portal_vendor_bills: Fix content in attachment files from portal
The use of method read() is not appropriated in this case, previous to
this change the read method was receiving, and empty bytes object since
stream was already at EOF. This behavior could be fixed with reset the
stream with seek() method.

But it's an clean solution to use the getvalue() method which Returns a
<type 'str'> containing the entire content of the buffer and the stream
position is not changed, unlike read() method.

You can check the '_io.BytesIO' documentation for different methods
involved read(), seek() and getvalue() here:
https://docs.python.org/3/library/io.html

Issue repoted on task 19453